// Tests/enums/product.ts
export enum ProductCategory {
    Women,
    Men,
    Unisex
}

export enum ProductColor {
    Blue = 'Blue',
    Black = 'Black',
    Red = 'Red',
    Green = 'Green',
    Yellow = 'Yellow'
}

export enum ProductSize {
    Small = 'Small',
    Medium = 'Medium',
    Large = 'Large',
    ExtraLarge = 'Extra Large'
}